refactored plugin_oar.py
